flask_backend/requirements.txt
Flask
Flask-CORS
requests
pyTigerGraph
reportlab
gunicorn
python-dotenv
typing_extensions
Faker